Time: 31 ms (93.40%), Space: 105.8 MB (45.76%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -7,7 +7,7 @@ public:
         for (; i < n; i++) {
             sum += nums[i];
             currentLength = i - left + 1;
-            if ((long long) currentLength * nums[i] - sum > k)  sum -= nums[left++];
+            if (currentLength * nums[i] - sum > k)  sum -= nums[left++];
         }
         
         return i - left;
